backend/requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.30.6
supabase==2.5.1
pydantic==2.9.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
